+++ 1.18/Common/src/main/java/fuzs/puzzleslib/api/client/core/v1/ClientModConstructor.java
@@ -10,6 +10,7 @@ import fuzs.puzzleslib.impl.PuzzlesLib;
 import fuzs.puzzleslib.impl.client.core.ClientFactories;
 import fuzs.puzzleslib.impl.core.ModContext;
 import net.minecraft.client.KeyMapping;
+import net.minecraft.client.Minecraft;
 import net.minecraft.client.color.block.BlockColor;
 import net.minecraft.client.color.item.ItemColor;
 import net.minecraft.client.model.geom.ModelLayerLocation;
@@ -122,7 +123,10 @@ public interface ClientModConstructor extends PairedModConstructor {
 
     /**
      * @param context add a search tree builder together with a token
+     *
+     * @deprecated replaced with direct access to the search registry via {@link ClientAbstractions#getSearchRegistry(Minecraft)}
      */
+    @Deprecated(forRemoval = true)
     default void onRegisterSearchTrees(final SearchRegistryContext context) {
 
     }