+++ 1.18/Fabric/src/main/java/fuzs/puzzleslib/impl/client/core/FabricClientAbstractions.java
@@ -1,12 +1,15 @@
 package fuzs.puzzleslib.impl.client.core;
 
 import fuzs.puzzleslib.api.client.core.v1.ClientAbstractions;
+import fuzs.puzzleslib.mixin.client.accessor.MinecraftFabricAccessor;
 import net.fabricmc.fabric.api.client.model.BakedModelManagerHelper;
 import net.fabricmc.fabric.api.client.rendering.v1.TooltipComponentCallback;
 import net.minecraft.client.KeyMapping;
+import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.screens.inventory.tooltip.ClientTooltipComponent;
 import net.minecraft.client.resources.model.BakedModel;
 import net.minecraft.client.resources.model.ModelManager;
+import net.minecraft.client.searchtree.SearchRegistry;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.inventory.tooltip.TooltipComponent;
 
@@ -27,5 +30,10 @@ public final class FabricClientAbstractions implements ClientAbstractions {
     @Override
     public BakedModel getBakedModel(ModelManager modelManager, ResourceLocation identifier) {
         return BakedModelManagerHelper.getModel(modelManager, identifier);
+    }
+
+    @Override
+    public SearchRegistry getSearchRegistry(Minecraft minecraft) {
+        return ((MinecraftFabricAccessor) minecraft).puzzleslib$getSearchRegistry();
     }
 }