+++ 1.18/Fabric/src/main/java/fuzs/puzzleslib/mixin/client/MinecraftFabricMixin.java
@@ -7,12 +7,14 @@ import fuzs.puzzleslib.api.event.v1.data.DefaultedValue;
 import fuzs.puzzleslib.mixin.client.accessor.KeyMappingFabricAccessor;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.Options;
+import net.minecraft.client.Timer;
 import net.minecraft.client.gui.screens.DeathScreen;
 import net.minecraft.client.gui.screens.Screen;
 import net.minecraft.client.gui.screens.TitleScreen;
 import net.minecraft.client.multiplayer.ClientLevel;
 import net.minecraft.client.multiplayer.MultiPlayerGameMode;
 import net.minecraft.client.player.LocalPlayer;
+import net.minecraft.client.renderer.GameRenderer;
 import net.minecraft.network.Connection;
 import net.minecraft.world.phys.BlockHitResult;
 import net.minecraft.world.phys.HitResult;
@@ -32,6 +34,16 @@ import java.util.Objects;
 
 @Mixin(Minecraft.class)
 public abstract class MinecraftFabricMixin {
+    @Shadow
+    @Final
+    public GameRenderer gameRenderer;
+    @Shadow
+    private boolean pause;
+    @Shadow
+    private float pausePartialTick;
+    @Shadow
+    @Final
+    private Timer timer;
     @Shadow
     @Nullable
     public ClientLevel level;
@@ -53,6 +65,16 @@ public abstract class MinecraftFabricMixin {
     private DefaultedValue<Screen> puzzleslib$newScreen;
     @Unique
     private boolean puzzleslib$attackCancelled;
+
+    @Inject(method = "runTick", at = @At(value = "INVOKE", target = "Lnet/minecraft/client/renderer/GameRenderer;render(FJZ)V", shift = At.Shift.BEFORE))
+    private void runTick$0(boolean renderLevel, CallbackInfo callback) {
+        FabricClientEvents.BEFORE_GAME_RENDER.invoker().onBeforeGameRender(Minecraft.class.cast(this), this.gameRenderer, this.pause ? this.pausePartialTick : this.timer.partialTick);
+    }
+
+    @Inject(method = "runTick", at = @At(value = "INVOKE", target = "Lnet/minecraft/client/renderer/GameRenderer;render(FJZ)V", shift = At.Shift.AFTER))
+    private void runTick$1(boolean renderLevel, CallbackInfo callback) {
+        FabricClientEvents.AFTER_GAME_RENDER.invoker().onAfterGameRender(Minecraft.class.cast(this), this.gameRenderer, this.pause ? this.pausePartialTick : this.timer.partialTick);
+    }
 
     @Inject(method = "setScreen", at = @At("HEAD"), cancellable = true)
     public void setScreen$0(@Nullable Screen newScreen, CallbackInfo callback) {